Add first.py
